Move nullability detection into its own package
Nullability detection sat in org.axonframework.common.annotation, which
misrepresents it: annotations are only one way a language expresses nullability,
and the Kotlin detector deliberately reads type metadata instead.

A subpackage of common.annotation was not an option. NO_CLASSES_SHOULD_DEPEND_
UPPER_PACKAGES is frozen for this module, and AnnotationBasedNullabilityDetector
needs AnnotationUtils from the parent package, so the dependency would have been
a new hierarchy violation. As a sibling package the same dependency is ordinary
and introduces no cycle, since common.annotation does not refer back.

The ServiceLoader registration files are renamed accordingly, as their name is
the service interface's fully qualified name.
