+++ common/src/main/resources/META-INF/services/org.axonframework.common.nullability.NullabilityDetector
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-org.axonframework.common.annotation.AnnotationBasedNullabilityDetector
+org.axonframework.common.nullability.AnnotationBasedNullabilityDetector
